basic vector functions

+++ include/geometry.h
@@ -2,15 +2,15 @@
 #define GEOMETRY_H
 
 typedef struct {
-	float x, y, z, w;
+	double x, y, z, w;
 } Vec4;
 
 typedef struct {
-	float x, y, z;
+	double x, y, z;
 } Vec3;
 
 typedef struct {
-	float x, y;
+	double x, y;
 } Vec2;
 
 #endif /* GEOMETRY_H */
